updating DDF functions
Restructured and some cluster support
